+++ packages/admin/admin-generator/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @comet/admin-generator
 
+## 8.14.0
+
+### Patch Changes
+
+- 540f8c8: Form: Use scope type from gql schema when using scope from context
+
+## 8.13.0
+
+### Minor Changes
+
+- 94abd9b: Add initialValuesAsProp config to generate a initialValues prop for the form that allows injecting dynamic initialValues
+
 ## 8.12.0
 
 ### Minor Changes